refactor(smarthome): refactor test code so it can be excluded from code coverage
to do it we need move all test code from mod files into own directory
