+++ smarthome/src/lib.rs
@@ -1,6 +1,6 @@
-// разрешим пока прототип
-#![allow(dead_code)]
-
 pub mod devices;
 pub mod location;
 pub mod providers;
+
+#[cfg(test)]
+mod tests;
