+++ smarthome/src/tests/devices/mod.rs
@@ -1,0 +1,2 @@
+mod socket;
+mod thermometer;
