+++ smarthome/src/tests/location/mod.rs
@@ -1,0 +1,2 @@
+mod home;
+mod room;
