+++ smarthome/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod devices;
+mod location;
